resolve bf16 error

+++ src/diffusers/models/attention_processor.py
@@ -85,7 +85,7 @@ class Attention(nn.Module):
         self._from_deprecated_attn_block = _from_deprecated_attn_block
 
         self.scale_qk = scale_qk
-        self.scale = dim_head**-0.5 if self.scale_qk else 1.0
+        self.scale = dim_head ** -0.5 if self.scale_qk else 1.0
 
         self.heads = heads
         # for slice_size > 0 the attention score computation
@@ -680,12 +680,12 @@ class CustomDiffusionAttnProcessor(nn.Module):
                 encoder_hidden_states = attn.norm_encoder_hidden_states(encoder_hidden_states)
 
         if self.train_kv:
-            key = self.to_k_custom_diffusion(
-                    encoder_hidden_states.to(self.to_k_custom_diffusion.weight.dtype)
-                    ).to(attn.to_q.weight.dtype)
-            value = self.to_v_custom_diffusion(
-                    encoder_hidden_states.to(self.to_v_custom_diffusion.weight.dtype)
-                    ).to(attn.to_q.weight.dtype)
+            key = self.to_k_custom_diffusion(encoder_hidden_states.to(self.to_k_custom_diffusion.weight.dtype)).to(
+                attn.to_q.weight.dtype
+            )
+            value = self.to_v_custom_diffusion(encoder_hidden_states.to(self.to_v_custom_diffusion.weight.dtype)).to(
+                attn.to_q.weight.dtype
+            )
         else:
             key = attn.to_k(encoder_hidden_states)
             value = attn.to_v(encoder_hidden_states)
@@ -1409,12 +1409,12 @@ class CustomDiffusionXFormersAttnProcessor(nn.Module):
                 encoder_hidden_states = attn.norm_encoder_hidden_states(encoder_hidden_states)
 
         if self.train_kv:
-            key = self.to_k_custom_diffusion(
-                    encoder_hidden_states.to(self.to_k_custom_diffusion.weight.dtype)
-                    ).to(attn.to_q.weight.dtype)
-            value = self.to_v_custom_diffusion(
-                    encoder_hidden_states.to(self.to_v_custom_diffusion.weight.dtype)
-                    ).to(attn.to_q.weight.dtype)
+            key = self.to_k_custom_diffusion(encoder_hidden_states.to(self.to_k_custom_diffusion.weight.dtype)).to(
+                attn.to_q.weight.dtype
+            )
+            value = self.to_v_custom_diffusion(encoder_hidden_states.to(self.to_v_custom_diffusion.weight.dtype)).to(
+                attn.to_q.weight.dtype
+            )
         else:
             key = attn.to_k(encoder_hidden_states)
             value = attn.to_v(encoder_hidden_states)
